+++ R/ie_expand.R
@@ -57,10 +57,11 @@ ie_expand <- function(lam) {
 }
 
 
-#' Compute w_j(t) = f_j(t) * prod_{i != j} F_i(t) for exponential components
+#' Component weight for exponential parallel systems
 #'
-#' For component j with rate \eqn{\lambda_j} and other rates
-#' \eqn{\lambda_{-j}}:
+#' Computes the contribution of component j to the parallel system
+#' density at time t. For component j with rate \eqn{\lambda_j} and
+#' other rates \eqn{\lambda_{-j}}:
 #' \deqn{w_j(t) = \lambda_j e^{-\lambda_j t}
 #'   \prod_{i \neq j} (1 - e^{-\lambda_i t})}
 #'